requirements-api.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.5.3

# Jinja2 for templates
jinja2==3.1.2

# ML models (optional, can be slow to install)
# Uncomment for NER functionality:
# torch>=2.2.0
# transformers>=4.36.0
